separated spi components
