+++ components/route/CMakeLists.txt
@@ -1,3 +1,3 @@
 idf_component_register(SRCS "route.c"
                     INCLUDE_DIRS "include"
-                    PRIV_REQUIRES spi)
+                    PRIV_REQUIRES spi_netif)
