+++ components/spi_internal/include/spi_internal.h
